nvim: migrate to lazy.nvim
