+++ nvim/init.lua
@@ -1,4 +1,1 @@
-require("user.plugins")
-require("user.settings")
-require("user.filetypes")
-require("user.auto")
+require("config")
