+++ nvim/lua/config/globals.lua
@@ -1,0 +1,3 @@
+vim.g.mapleader = ","
+vim.g.maplocalleader  = ","
+vim.g.loaded_netrwPlugin = 0
