+++ nvim/lua/config/init.lua
@@ -1,0 +1,6 @@
+require("config.globals")
+require("config.autocmds")
+require("config.filetypes")
+require("config.options")
+require("config.keymaps")
+require("config.lazy")
