+++ nvim/lua/config/options.lua
@@ -1,0 +1,54 @@
+vim.opt.path = vim.opt.path + ",**"
+vim.opt.compatible = false
+vim.opt.number = true
+vim.opt.relativenumber = true
+vim.opt.cursorline = true
+vim.opt.cursorcolumn = false
+vim.opt.hidden = true
+vim.opt.magic = true
+vim.opt.showmatch = true
+vim.opt.autowrite = true
+vim.opt.hlsearch = true
+vim.opt.ignorecase = true
+vim.opt.smartcase = true
+vim.opt.colorcolumn = "80"
+vim.opt.termguicolors = true
+vim.opt.updatetime = 50
+vim.opt.history = 10000
+vim.opt.autoread = true
+vim.opt.timeout = false
+vim.opt.ttimeout = true
+vim.opt.ttimeoutlen = 0
+vim.opt.ttyfast = true
+vim.opt.signcolumn = "auto"
+vim.opt.backup = false
+vim.opt.swapfile = false
+vim.opt.wrap = false
+vim.opt.textwidth = 0
+vim.opt.eol = false
+vim.opt.list = true
+vim.opt.listchars = {
+    tab = "▸ ",
+    trail = "•",
+    extends = "»",
+    precedes = "«",
+}
+vim.opt.scrolloff = 4
+vim.opt.laststatus = 2
+vim.opt.lazyredraw = true
+vim.opt.showmode = false
+vim.opt.completeopt = "menu,menuone,noselect"
+vim.opt.switchbuf = "usetab"
+vim.opt.undofile = true
+vim.opt.undodir = os.getenv("HOME") .. "/.local/share/nvim/undodir"
+vim.opt.tm = 500
+vim.opt.maxmempattern = 20000
+vim.opt.inccommand = "split"
+vim.opt.autoindent = true
+vim.opt.cindent = true
+vim.opt.tabstop = 4
+vim.opt.shiftwidth = 4
+vim.opt.softtabstop = 4
+vim.opt.expandtab = true
+vim.opt.wildmode = "longest:full"
+vim.opt.diffopt:append("linematch:60")
