+++ nvim/lua/plugins/go-nvim.lua
@@ -1,0 +1,4 @@
+return {
+  "ray-x/go.nvim",
+  config = true,
+}
