+++ nvim/lua/plugins/null-ls.lua
@@ -1,0 +1,18 @@
+return {
+  "jose-elias-alvarez/null-ls.nvim",
+  event = "BufReadPre",
+  opts = function()
+    local null_ls = require("null-ls")
+    return {
+      sources = {
+        null_ls.builtins.formatting.prettier,
+        null_ls.builtins.formatting.eslint_d,
+        null_ls.builtins.diagnostics.eslint_d,
+        null_ls.builtins.formatting.stylua,
+      },
+      should_attach = function(bufnr)
+        return not vim.api.nvim_buf_get_name(bufnr):match("^git://")
+      end,
+    }
+  end,
+}
