+++ nvim/lua/plugins/plenary.lua
@@ -1,0 +1,4 @@
+return {
+  "nvim-lua/plenary.nvim",
+  lazy = true,
+}
