+++ nvim/lua/plugins/splitjoin.lua
@@ -1,0 +1,3 @@
+return {
+  "AndrewRadev/splitjoin.vim",
+}
