+++ nvim/lua/plugins/undotree.lua
@@ -1,0 +1,7 @@
+return {
+  "mbbill/undotree",
+  cmd = "UndotreeToggle",
+  keys = {
+    { "<leader>u", vim.cmd.UndotreeToggle, noremap = true, silent = true },
+  },
+}
