+++ nvim/lua/plugins/vim.lua
@@ -1,0 +1,34 @@
+return {
+  {
+    "tpope/vim-dispatch",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-repeat",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-sleuth",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-surround",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-unimpaired",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-eunuch",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-commentary",
+    event = "BufEnter",
+  },
+  {
+    "tpope/vim-git",
+    event = "BufEnter",
+  },
+}
