--- nvim/lua/user/configs/dirvish.lua
@@ -1,6 +1,0 @@
-vim.cmd([[
-  augroup dirvish_config
-      autocmd!
-      autocmd FileType dirvish silent! unmap <buffer> <C-p>
-  augroup end
-]])
