--- nvim/lua/user/configs/go-nvim.lua
@@ -1,1 +1,0 @@
-require("go").setup()
