+++ nvim/stylua.toml
@@ -1,6 +1,6 @@
-column_width = 120
-line_endings = "Unix"
 indent_type = "Spaces"
-indent_width = 4
+indent_width = 2
+column_width = 120
 quote_style = "AutoPreferDouble"
 call_parentheses = "Always"
+line_endings = "Unix"
